q1_data_cleaning/q1_requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
cloudstorage==0.11.0
decorator==5.1.1
frozenlist==1.3.0
fsspec==2022.2.0
gcsfs==2022.2.0
google-api-core==2.5.0
google-auth==2.6.0
google-auth-oauthlib==0.5.0
google-cloud-bigquery==2.34.0
google-cloud-core==2.2.2
google-cloud-storage==2.1.0
google-crc32c==1.3.0
google-resumable-media==2.3.0
googleapis-common-protos==1.55.0
grpcio==1.44.0
grpcio-status==1.44.0
idna==3.3
inflection==0.5.1
multidict==6.0.2
numpy==1.21.5
oauthlib==3.2.0
packaging==21.3
pandas==1.3.5
proto-plus==1.20.3
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.7
python-dateutil==2.8.2
python-magic==0.4.25
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
typing_extensions==4.1.1
urllib3==1.26.8
yarl==1.7.2
